refactor(tests): move loop_agent fixture to tests/run_agent/conftest.py
This exact fixture body was already duplicated verbatim across several
files in tests/run_agent/ (test_dropped_tool_call_recovery.py and others)
before this PR added an Nth copy in test_fabricated_tool_use_recovery.py.
conftest.py in this directory already exists as the designated home for
shared fixtures (it holds the autouse _fast_retry_backoff fixture) but
held no loop_agent.

Scoped to just the copy this PR itself contributed -- the pre-existing
duplication across test_dropped_tool_call_recovery.py and ~5 other files
is a separate, larger cleanup not attempted here.

Found via an independent post-merge review of the PR.
